+++ clang/lib/DPCT/CMakeLists.txt
@@ -156,6 +156,7 @@ add_clang_library(DPCT
   IncrementalMigrationUtility.cpp
   Rules.cpp
   PatternRewriter.cpp
+  MigrateCmakeScript.cpp
   Homoglyph.cpp
   MisleadingBidirectional.cpp
   BarrierFenceSpaceAnalyzer.cpp
